requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
opencv-python==4.1.1.26
Pillow==6.2.1
python-resize-image==1.1.19
requests==2.22.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
